requirements.txt
pytorch
torchvision
numpy
tqdm
scikit-learn
pandas
matplotlib
